// frontend/src/api/generated/models/routingPayload.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface RoutingPayload {
  product_model_id: string;
  version?: number;
  effective_date: string;
  is_active?: boolean;
}